requirements.txt
pyyaml
git+https://github.com/Pycord-Development/pycord
openai==0.27.0